feat: Tratamento de exceção no serviço de deleção de usuario.

+++ src/main/java/com/gabriel/course/services/UserService.java
@@ -2,8 +2,11 @@ package com.gabriel.course.services;
 
 import com.gabriel.course.entities.User;
 import com.gabriel.course.repositories.UserRepository;
+import com.gabriel.course.services.exceptions.DatabaseException;
 import com.gabriel.course.services.exceptions.ResourceNotFoundException;
+import org.apache.tomcat.util.buf.UDecoder;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -29,7 +32,15 @@ public class UserService {
     }
 
     public void delete(Long id) {
-        repository.deleteById(id);
+        if(!repository.existsById(id)){
+            throw new ResourceNotFoundException(id);
+        }
+        try {
+            repository.deleteById(id);
+        }
+        catch (DataIntegrityViolationException e) {
+            throw new DatabaseException(e.getMessage());
+        }
     }
 
     public User update(Long id, User obj) {
